requirements.txt
Flask
Flask-Login
Flask-Mail
Flask-SQLAlchemy
gitpython
cython
mistune
pygments
itsdangerous
Pillow
unidiff
